Add files via upload
Okay, first of all. Hi, everyone who travel this website. So, my name is Shivam Sharma... So, ya now we talk about this website and it goes like that I made this websites for college project for CHE110(it's subject code of my subject) and in which I use basic html and CSS. And in this website I built three basic pages and in first page there is nav bar on that there is only logo there. And next is the landing page in which I use video for better experience and next I use some picture which are clickable and in them i used Wikipedia link for reference and next is like overview of these project that are run by Indian government or some private also and next is our mission and vision and then a footer portion.. second page is just for some projects that are successfully run by the govt. and privately also with some link.. third page is goes like that in that also some other project of conservation of biodiversity .. that's it.. thanks for give your precious time here.